move timestamp code from lib to mod
